+++ thongssh_gtk/constants.py
@@ -1,7 +1,7 @@
 import sys
 import os
 
-__version__ = "0.6.1"
+__version__ = "0.6.2"
 
 APP_ID = "com.example.thongssh"